app/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
databricks-sql-connector>=3.0.0
pydantic>=2.5.0
